fix(platform): per-scrape NonFatal + bucketCount Long — review hardening follow-ups
Two follow-ups from the freshness-review pass that were not picked up
by the TTL commit's git add (oversight in the staging step):

- RollupFreshnessReader: drop the .toInt truncation on bucketCount;
  the Entry field is already Long. Watermark tables easily exceed
  Int.MaxValue over multi-year retention (the reviewer's MEDIUM #2).

- Main.installRollupFreshnessReader: wrap the per-scrape reflective
  call in NonFatal. The install-time catch only covers boot wiring
  failures; a runtime failure mid-scrape (connector upgrade renaming
  the reflective target, InvocationTargetException from readAll) must
  NOT escape into the Vert.x handler thread and 500 /metrics for
  every verb. Degrade to a typed Left — the platform's probe_failed
  gauge is the signal.

+++ connectors/spark-connector/src/main/scala/io/sm8/connectors/spark/RollupFreshnessReader.scala
@@ -33,8 +33,10 @@ final case class FreshnessEntry(
   /** Aggregate finality: true iff EVERY bucket row is is_final=true.
     * A rollup with zero rows reports false (nothing proven final). */
   allFinal:        Boolean,
-  /** Number of bucket rows in the watermark table. */
-  bucketCount:     Int
+  /** Number of bucket rows in the watermark table. `Long` to
+    * accommodate multi-year retention (Int.MaxValue ≈ 24-day
+    * hourly buckets; few-years daily buckets easily exceed Int). */
+  bucketCount:     Long
 )
 
 object RollupFreshnessReader {
@@ -85,7 +87,7 @@ object RollupFreshnessReader {
               case b: java.lang.Boolean => b.booleanValue()
               case _ => false
             },
-            bucketCount     = bucketCount.toInt
+            bucketCount     = bucketCount
           )
         }
         attempted.toEither match {
